Add top5 endpoints for events and serviceproducts

+++ src/main/java/com/example/eventplanner/repositories/user/UserReportRepository.java
@@ -3,5 +3,5 @@ package com.example.eventplanner.repositories.user;
 import com.example.eventplanner.model.user.UserReport;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserReportRepository extends JpaRepository<UserReport, Integer> {
+public interface UserReportRepository extends JpaRepository<UserReport, Long> {
 }
